myenv/requirements.txt
streamlit 
faiss-cpu 
sentence-transformers 
pypdf 
numpy 
openai
langchain